Major code cleanup
This commit is a major cleanup of most of the job and terminal handling
which should make the code easier to follow and maintain.

I've also attempted to limit how invasive the terminal window is to your
current window layout. It works for me, and I hope it works for other
workflows as well.